+++ Mytube/src/main/java/com/flab/Mytube/service/ConvertMovieService.java
@@ -1,7 +1,6 @@
 package com.flab.Mytube.service;
 
 import com.flab.Mytube.domain.Movie;
-import com.flab.Mytube.dto.movie.request.ChuncksBuildRequest;
 import com.flab.Mytube.dto.movie.request.FileUploadRequest;
 import com.flab.Mytube.dto.movie.request.MovieDtailRequest;
 import com.flab.Mytube.error.exceptions.NoDataSubmitException;
@@ -13,11 +12,6 @@ import com.flab.Mytube.utils.Validations;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
-import net.bramp.ffmpeg.FFmpeg;
-import net.bramp.ffmpeg.FFmpegExecutor;
-import net.bramp.ffmpeg.FFprobe;
-import net.bramp.ffmpeg.builder.FFmpegBuilder;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
@@ -35,12 +29,10 @@ import java.nio.file.Path;
 public class ConvertMovieService {
 
   private final MovieMapper movieMapper;
-  private final FFmpeg fFmpeg;
-  private final FFprobe fFprobe;
   private final MoviePath moviePath;
 
   @Value("${kafka.encoding.topic}")
-  String TOPIC ;
+  String TOPIC;
 
   @Autowired
   Producer producer;
@@ -65,40 +57,10 @@ public class ConvertMovieService {
     String data = originPath.toString();
 
     String key = fileName.split("\\.")[0];
-//    producer.send(data);
+    File chunckPath = moviePath.chunckPath(data);
     producer.send(TOPIC, key, data);
 //    request.addPath(MoviePath.chunkPathStr(originPath.toString()));
 //    movieMapper.save(request);
-  }
-
-
-//  @KafkaListener(topics = "videoPath", groupId = "myGroup", containerFactory = "kafkaListenerContainerFactory")
-  public void segment(ConsumerRecord<String, String> data) {
-    String originPath = data.value();
-    File chunckPath = moviePath.chunckPath(originPath);
-    String fileName = chunckPath.getName().split("\\.")[0];
-
-    ChuncksBuildRequest chunkBuilder = ChuncksBuildRequest.builder()
-        .name(fileName)
-        .originPath(originPath)
-        .m3u8Name(fileName + ".m3u8")
-        .m3u8Path(chunckPath)
-        .build();
-    FFmpegBuilder builder = Movies.segmentationTs(chunkBuilder);
-
-    try {
-      run(builder);
-    } catch (IllegalArgumentException e) {
-      log.info("N/A error ocuuer");
-    } catch (Exception e) {
-      log.info("영상 변환 중 에러가 발생했습니다. 다시 시도해주세요.");
-    }
-  }
-
-  private void run(FFmpegBuilder builder) throws Exception {
-    FFmpegExecutor executor = new FFmpegExecutor(fFmpeg, fFprobe);
-
-    executor.createJob(builder).run();
   }
